Fix int overflow in recommended army/navy size computation
UBSan finding 7 (CvMilitaryAI.cpp:1820):
'217 * 9974000 cannot be represented in int' in
SetRecommendedArmyNavySize.

The defense/offense weights are scaled by per-city contributions and
then multiplied by iDefenseModifier/iOffenseModifier (~100-200), so a
large empire legitimately reaches 5-digit weights (locals:
iLandDefenseWeight=70760, iTotalOffenseWeight=34500, 33 cities). The
numerator 'iMaxPossibleUnits * (100 * weight + offense * percent)'
then exceeds INT_MAX (217 * 9,974,000 = 2.16e9). The overflow entered
with commit 605684b26 which multiplied the unit cap into the
weighted-share formula. The sibling naval line has the identical
structure and overflows the same way.

Widen the integer numerator to long long before the float division.
The division and rounding stay in float as before, so results are
unchanged for all non-overflowing inputs.

+++ CvGameCoreDLL_Expansion2/CvMilitaryAI.cpp
@@ -1817,8 +1817,9 @@ void CvMilitaryAI::SetRecommendedArmyNavySize()
 	// If we have more explorers than we need, don't build more units above the force limit anyway
 	iMaxPossibleUnits -= max(iExplorersNeeded, iNumExplorers);
 	
-	m_iRecLandUnits = static_cast<int>((iMaxPossibleUnits * (100 * iLandDefenseWeight + iTotalOffenseWeight * (100 - iNavalPercent)) / (100.0f * iTotalWeight)) + 0.5f);
-	m_iRecNavalUnits = static_cast<int>((iMaxPossibleUnits * (100 * iNavalDefenseWeight + iTotalOffenseWeight * iNavalPercent)) / (100.0f * iTotalWeight) + 0.5f);
+	// 64 bit math for the numerator - the modifier-scaled weights times the unit cap exceed INT_MAX for large late-game empires
+	m_iRecLandUnits = static_cast<int>(((long long)iMaxPossibleUnits * (100LL * iLandDefenseWeight + (long long)iTotalOffenseWeight * (100 - iNavalPercent)) / (100.0f * iTotalWeight)) + 0.5f);
+	m_iRecNavalUnits = static_cast<int>(((long long)iMaxPossibleUnits * (100LL * iNavalDefenseWeight + (long long)iTotalOffenseWeight * iNavalPercent)) / (100.0f * iTotalWeight) + 0.5f);
 
 	int iNumLandUnits = m_pPlayer->GetNumUnitsWithDomain(DOMAIN_LAND, true) - iNumExplorers;
 	int iNumNavalUnits = m_pPlayer->GetNumUnitsWithDomain(DOMAIN_SEA, true);
